Move REST API portion of plugin to its own class
